--- bard/util/query.py
@@ -1,7 +1,0 @@
-'''
-name: property ({name: 'property'})
-name: 'lol property: xd' ({name: 'lol property: xd'})
-name: *Last Week Tonight*
-id: 1 ({id: 1})
-'''
-
