--- config.example.json
@@ -1,34 +1,0 @@
-{
-  "providers": {
-    "info": {"name": "tvdb"},
-    "download": {"name": "iptorrents", "username": "gadget", "password": "gadget"},
-    "fetch": {"name": "transmission", "url": "https://transmission.hydr0.com", "username": "gadget", "password": "gadget"},
-    "notify": {"name": "discord", "url": "https://discordapp.com/api/webhooks/numbers_here/lots_of_text_here"},
-    "library": {
-      "name": "plex",
-      "url": "https://yo-plex-is-pretty-cool",
-      "token": "token bro"
-    }
-  },
-  "directories": {
-    "input": "/media/transmission",
-    "output": "/media/storage/media/TV",
-    "temporary": "/media/storage/media/Staging"
-  },
-  "web": {
-    "secret_key": "hi :)",
-    "host": "",
-    "port": 7675
-  },
-  "default_quality": "1080p",
-  "acls": {},
-  "info": {
-    "providers": [{"name": "tvdb"}, {"name": "tvmaze"}],
-    "search": "tvdb",
-    "rules": {
-      "series": {"*": "tvdb"},
-      "season": {"*": "tvdb"},
-      "episode": {"*": "tvdb", "airdate": "tvmaze"}
-    }
-  }
-}
